+++ Frogalicious/Assets/Modules/Level/Code/Primitives/BoardTileType.cs
@@ -1,6 +1,6 @@
 namespace Frog.Level.Primitives
 {
-    public enum BoardTileType : uint
+    public enum BoardTileType : ushort
     {
         Nothing = 0,
         Ground = 1,
